refactor: Move serializeFeatureAttributes fucntion
From HsQueryBaseService to HsQueryPopupBaseService
