+++ projects/hslayers/src/components/query/query-base.service.ts
@@ -18,7 +18,6 @@ import {HsLayoutService} from '../layout/layout.service';
 import {HsMapService} from '../map/map.service';
 import {HsSaveMapService} from '../save-map/save-map.service';
 import {HsUtilsService} from '../utils/utils.service';
-import {getPopUp} from '../../common/layer-extensions';
 
 @Injectable({
   providedIn: 'root',
@@ -123,52 +122,6 @@ export class HsQueryBaseService {
         const layer = this.HsMapService.getLayerForFeature(feature);
         return layer && layer != this.queryLayer;
       });
-  }
-
-  /**
-   * @param feature
-   */
-  serializeFeatureAttributes(feature: Feature<Geometry>): any[] {
-    const attributesForHover = [];
-    const layer = this.HsMapService.getLayerForFeature(feature);
-    if (layer === undefined) {
-      return;
-    }
-    let attrsConfig = [];
-    if (getPopUp(layer)?.attributes) {
-      //must be an array
-      attrsConfig = getPopUp(layer).attributes;
-    } else {
-      // Layer is not configured to show pop-ups
-      return;
-    }
-    for (const attr of attrsConfig) {
-      let attrName, attrLabel;
-      let attrFunction = (x) => x;
-      if (typeof attr === 'string' || attr instanceof String) {
-        //simple case when only attribute name is provided in the layer config
-        attrName = attr;
-        attrLabel = attr;
-      } else {
-        if (attr.attribute == undefined) {
-          //implies malformed layer config - 'attribute' is obligatory in this case
-          continue;
-        }
-        attrName = attr.attribute;
-        attrLabel = attr.label != undefined ? attr.label : attr.attribute;
-        if (attr.displayFunction) {
-          attrFunction = attr.displayFunction;
-        }
-      }
-      if (feature.get(attrName)) {
-        attributesForHover.push({
-          key: attrLabel,
-          value: feature.get(attrName),
-          displayFunction: attrFunction,
-        });
-      }
-    }
-    return attributesForHover;
   }
 
   setData(data: any, type: string, overwrite?: boolean): void {
