+++ projects/hslayers/src/components/query/query-popup-base.service.ts
@@ -5,7 +5,7 @@ import {Geometry} from 'ol/geom';
 
 import {HsMapService} from '../map/map.service';
 import {HsUtilsService} from '../utils/utils.service';
-import {getTitle} from '../../common/layer-extensions';
+import {getPopUp, getTitle} from '../../common/layer-extensions';
 
 @Injectable({
   providedIn: 'root',
@@ -50,5 +50,51 @@ export class HsQueryPopupBaseService {
 
   closePopup(): void {
     this.featuresUnderMouse = [];
+  }
+
+  /**
+   * @param feature -
+   */
+  serializeFeatureAttributes(feature: Feature<Geometry>): any[] {
+    const attributesForHover = [];
+    const layer = this.HsMapService.getLayerForFeature(feature);
+    if (layer === undefined) {
+      return;
+    }
+    let attrsConfig = [];
+    if (getPopUp(layer)?.attributes) {
+      //must be an array
+      attrsConfig = getPopUp(layer).attributes;
+    } else {
+      // Layer is not configured to show pop-ups
+      return;
+    }
+    for (const attr of attrsConfig) {
+      let attrName, attrLabel;
+      let attrFunction = (x) => x;
+      if (typeof attr === 'string' || attr instanceof String) {
+        //simple case when only attribute name is provided in the layer config
+        attrName = attr;
+        attrLabel = attr;
+      } else {
+        if (attr.attribute == undefined) {
+          //implies malformed layer config - 'attribute' is obligatory in this case
+          continue;
+        }
+        attrName = attr.attribute;
+        attrLabel = attr.label != undefined ? attr.label : attr.attribute;
+        if (attr.displayFunction) {
+          attrFunction = attr.displayFunction;
+        }
+      }
+      if (feature.get(attrName)) {
+        attributesForHover.push({
+          key: attrLabel,
+          value: feature.get(attrName),
+          displayFunction: attrFunction,
+        });
+      }
+    }
+    return attributesForHover;
   }
 }
